"--daemon" option not working properly ("./installer command not found").
Why? Because we are calling "installer" with "&" which makes it to run in background and return immediately.
After return we destroy its tmpdir, including the main "installer" file.

Solution: Remove tmpdir at end of the "installer" script. Also change "exec" to nothing, because nothing after "exec" will get run.
Exec replaces the process image of the current process with the process image of the executable you exec.
That means that the moment you exec, the shell process that does the execing gets completely destroyed and replaced by the execed program.